api/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
furl==2.1.3
gunicorn==20.1.0
orderedmultidict==1.0.1
psycopg2-binary==2.9.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
